requirements.txt
# sudo apt-get install libjpeg-dev

pillow
